Reshaping of the project structure and forcing compliance with some good practices
