--- src/lexical.cpp
@@ -1,1 +1,0 @@
-#include <lexical.hpp>
